// client/src/interfaces/City.ts
export interface City {
  postalCode: string;
  City: string;
  State: string;
  Region: string;
  NeighborhoodType: string;
  Diversity: string;
  Politics: string;
  ageTwentyTwoToThirtyFour: string;
  ageThirtyFiveToFiftyFour: string;
  ageFiftyFiveToSeventyFour: string;
  twoByThreeToFourByFiveBed: string;
  twoByThreeBed: string;
  fourByFiveBed: string;
  bestSchool: string;
}